requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dateparser==1.2.0
deepmerge==1.1.1
dnspython==2.6.1
fastapi==0.110.1
flatdict==4.0.1
h11==0.14.0
idna==3.7
loguru==0.7.2
marshmallow==3.21.1
marshmallow-mongoengine==0.31.2
mongoengine==0.28.2
packaging==24.0
passlib==1.7.4
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==1.7.1
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
regex==2023.12.25
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
tzlocal==5.2
uvicorn==0.29.0
